test macOS keychain creds, remove bash scripts in favor of spawning child_processes with stdin args in node

+++ app.js
@@ -143,33 +143,39 @@ utils.getIPForDisplay().then(data=>{
   }
   
   console.log("HandyHost Daemon Running at: http://"+data.ip+":" + data.port + "/, and https://"+data.ip+":"+httpsPort+'/ (self-signed cert)');
-  console.log("SCAUTO ENV",process.env.SCAUTO)
+  /*console.log("SCAUTO ENV",process.env.SCAUTO)
   console.log("AKTAUTO ENV",process.env.AKTAUTO)
-  console.log("DVPNAUTO ENV",process.env.DVPNAUTO)
+  console.log("DVPNAUTO ENV",process.env.DVPNAUTO)*/
 })
 
 
 process.on('uncaughtException', function(err) {
-  if(err.code.indexOf('EADDRINUSE') >= 0){
-    utils.getIPForDisplay().then(data=>{
-      if(process.platform == 'darwin'){
-        const startupLog = process.env.HOME+'/.HandyHost/startup.log';
-        const line0 = "NOTIFICATION: HandyHost Daemon Already Running at: http://"+data.ip+":" + data.port + "/ and https://"+data.ip+":"+httpsPort+' (self-signed cert)';
-        const line0a = line0.replace('NOTIFICATION: ','');
-        const line1 = 'To stop the daemon, open a terminal and run: ';
-        const line2 = 'kill '+fs.readFileSync(process.env.HOME+'/.HandyHost/handyhost.pid','utf8')
-        fs.writeFileSync(startupLog,line0+'\n','utf8'); //will fire a notification
-        fs.appendFileSync(startupLog,line0a+'\n','utf8'); //will log to the UI log panel
-        fs.appendFileSync(startupLog,line1+'\n','utf8');
-        fs.appendFileSync(startupLog,line2+'\n','utf8');  
-        console.log(line0);
-        console.log(line0a);
-        console.log(line1);
-        console.log(line2);
-      }
-      console.log("HandyHost Daemon Already Running at: http://"+data.ip+":" + data.port + "/ and https://"+data.ip+":"+httpsPort+' (self-signed cert)');
+  if(typeof err.code != "undefined"){
+    if(err.code.indexOf('EADDRINUSE') >= 0){
+      utils.getIPForDisplay().then(data=>{
+        if(process.platform == 'darwin'){
+          const startupLog = process.env.HOME+'/.HandyHost/startup.log';
+          const line0 = "NOTIFICATION: HandyHost Daemon Already Running at: http://"+data.ip+":" + data.port + "/ and https://"+data.ip+":"+httpsPort+' (self-signed cert)';
+          const line0a = line0.replace('NOTIFICATION: ','');
+          const line1 = 'To stop the daemon, open a terminal and run: ';
+          const line2 = 'kill '+fs.readFileSync(process.env.HOME+'/.HandyHost/handyhost.pid','utf8')
+          fs.writeFileSync(startupLog,line0+'\n','utf8'); //will fire a notification
+          fs.appendFileSync(startupLog,line0a+'\n','utf8'); //will log to the UI log panel
+          fs.appendFileSync(startupLog,line1+'\n','utf8');
+          fs.appendFileSync(startupLog,line2+'\n','utf8');  
+          console.log(line0);
+          console.log(line0a);
+          console.log(line1);
+          console.log(line2);
+        }
+        console.log("HandyHost Daemon Already Running at: http://"+data.ip+":" + data.port + "/ and https://"+data.ip+":"+httpsPort+' (self-signed cert)');
+        process.exit(1);
+      })
+    }
+    else{
+      console.log('Caught exception: ' + err);
       process.exit(1);
-    })
+    }
   }
   else{
     console.log('Caught exception: ' + err);
